feat: Profileimage 공통 컴포넌트 추가
- components/common/profile/ProfileImage.tsx 추가
  - props로 width, height 전달

Related to: #174
